Add react testing library setup and convert 1 test
