+++ package.json
@@ -37,7 +37,9 @@
         "fix:tslint": "tslint --fix --project .",
         "test": "run-s test:e2e",
         "test:lint": "tslint --project . && prettier \"src/**/*.ts\" \"src/**/*.tsx\" \"test/**/*.ts\" \"test/**/*.tsx\" \"./*.json\" --list-different",
-        "test:unit": "jest --env=node --no-cache --colors --coverage test/unit/specs --setupFilesAfterEnv ./test/unit/setup.js",
+        "test:unit": "run-s test:unit:*",
+        "test:unit:enzyme": "jest --env=node --no-cache --colors --coverage test/unit/specs --setupFilesAfterEnv ./test/unit/setup.js",
+        "test:unit:rtl": "jest src/ --config ./test/unit-rtl/jest.config.js",
         "test:e2e": "run-s test:e2e:*",
         "test:e2e:react-14": "cross-env REACT_VERSION=0.14.0 run-s set-react-14 dotenv build:e2e clean-react-alias run-jest",
         "test:e2e:react-15": "cross-env REACT_VERSION=15.0.0 run-s set-react-15 dotenv build:e2e clean-react-alias run-jest",
@@ -59,6 +61,8 @@
     },
     "devDependencies": {
         "@fluffy-spoon/substitute": "^1.89.0",
+        "@testing-library/jest-dom": "^5.16.5",
+        "@testing-library/react": "^12.1.5",
         "@types/enzyme": "^3.9.1",
         "@types/expect-puppeteer": "^3.3.1",
         "@types/jest": "^24.0.11",
@@ -73,6 +77,7 @@
         "express": "^4.16.4",
         "husky": "^1.3.1",
         "jest": "^24.6.0",
+        "jest-environment-jsdom-fifteen": "^1.0.2",
         "jest-pnp-resolver": "^1.2.1",
         "jest-puppeteer": "^4.1.0",
         "jest-resolve": "^24.6.0",
